Practice cascade in css
In this exercise, I praticed how to apply orders in styling in css using cascade. It is mostly about how differnt rule gets precedence over other based on the importance of the order in which they come in, which is defined by their specifity. So I learn how to use the specifity by either using the most specific selectors or chaining certains selectors together to have a certain priority in application of the rules